WIP Add OS::Cinder::Quota resource
This creates new resource type whose intended use case is for admin-only
use and specifically to manage the cinder quota. This would be an enhancement
to avoid manual, post-deployment modifications of the projects cinder quota.
We want to enable the admin to fully orchestrate the project by using heat
without the need to make manual updates afterwards.

Implements bp/cinder-quota-resource

Change-Id: I49d01d229199d9c472dc59ba2bb95d455f6dfb76